requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
websockets==12.0
python-dotenv==1.0.1
pydantic==2.5.3
pydantic-settings==2.2.1
openai==1.12.0
twilio==9.1.3
httpx==0.27.0
SQLAlchemy==2.0.25
alembic==1.13.1
python-jose==3.3.0
